requirements.txt
requests==2.13.0
python-dateutil==2.6.0
mock==2.0.0
pytest==3.0.6
